added error handling, cli, gui is now its own file
