+++ Cargo.toml
@@ -4,6 +4,8 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
+clap = { version = "4.5.39", features = ["derive"] }
+color-eyre = "0.6.5"
 iced = "0.13.1"
 iced_layershell = "0.13.6"
 iced_runtime = "0.13.2"
