+++ Justfile
@@ -1,0 +1,6 @@
+
+default:
+  @just --list
+
+run:
+  cargo run -- -i bindings.json
